More functionality and restrictions on users not logged in
I now store who uploads images, likes, and comments. Currently, I show
usernames with photos and comments.

I also hid some features for users who are not logged in, so that they
cannot upload an image, like, comment, delete (has to match who
uploaded), etc. So people who are not logged in shouldn’t be able to
tamper with it.

Also, since the upload image menu is at the bottom of the gallery, I
added a link for users who are logged in to jump to that section.